proactive specs

+++ typescript-plugin-proactive-code-fixes/spec/testUtil.ts
@@ -1,24 +1,35 @@
 import * as shell from 'shelljs';
-import Project, { TypeGuards, Program, SourceFile } from 'ts-simple-ast';
+import Project, { TypeGuards, Program, SourceFile, ScriptTarget, ModuleKind } from 'ts-simple-ast';
 import { CodeFix } from '../src/typings';
 
 export function defaultBeforeEach(config: DefaultBeforeEachInput): DefaultBeforeEachResult {
-  shell.rm(`-rf`, config.projectPath);
-  shell.cp(`-r`, `assets/sampleProject1`, `${config.projectPath}`);
-  const simpleProject = new Project({
-    tsConfigFilePath: `${config.projectPath}/tsconfig.json`
-  });
+  let simpleProject: Project
+  if(config.projectPath){
+    shell.rm(`-rf`, config.projectPath);
+    shell.cp(`-r`, `assets/sampleProject1`, `${config.projectPath}`);
+     simpleProject = new Project({
+      tsConfigFilePath: `${config.projectPath}/tsconfig.json`
+    });
+  }
+  else {
+    simpleProject = new Project({
+      compilerOptions: {
+          target: ScriptTarget.ES2018,
+          lib: ["es2018"], module: ModuleKind.CommonJS
+      }
+    })
+  }
   const program = simpleProject.getProgram().compilerObject
   let newSourceFile: SourceFile
   if (config.createNewFile) {
-    newSourceFile = simpleProject.createSourceFile('tmp/sourceFileTmp__' + Date.now() + '.ts', '')
+    newSourceFile = simpleProject.createSourceFile('sourceFileTmp__' + Date.now() + '.ts', typeof config.createNewFile==='string' ? config.createNewFile : '')
   }
   return Object.assign({}, config, { program, simpleProject, newSourceFile })
 }
 
 export interface DefaultBeforeEachResult extends DefaultBeforeEachInput { program: ts.Program, simpleProject: Project, newSourceFile: SourceFile | undefined }
 
-export interface DefaultBeforeEachInput{ projectPath: string, createNewFile?: boolean }
+export interface DefaultBeforeEachInput{ projectPath?: string, createNewFile?: boolean|string }
 export function expectToContainFixer(fixes: CodeFix[], name: string): CodeFix|undefined {
   let fix
   if ((fix = fixes.find(f => f.name == name))) {
